% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{deconv_performance_plot}
\alias{deconv_performance_plot}
\title{Title}
\usage{
deconv_performance_plot(
  est,
  true,
  title = NULL,
  nrow = 2,
  xlabel = "true fraction",
  ylabel = "estimates"
)
}
\arguments{
\item{est}{a}

\item{true}{a}

\item{title}{a}

\item{nrow}{a}
}
\description{
Title
}
