% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{fastPost.ini.cs.cpp}
\alias{fastPost.ini.cs.cpp}
\title{Fast posterior estimation of cell-state fractions and expression}
\usage{
fastPost.ini.cs.cpp(bulk_Expr, ref, n.iter, n.core = 1)
}
\arguments{
\item{bulk_Expr}{bulk expression (un-log transformed) matrix with genes in rows and samples in columns}

\item{ref}{single cell reference summarized at per cell-type or per cell-state level}

\item{n.iter}{number of iterations.}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\description{
Fixed-point implementation of the initial Gibbs sampling step in BayesPrism.
     Returns a list of posterior estimation of cell-state fraction and expression.
}
