% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{fastPost.updated.ct.cpp}
\alias{fastPost.updated.ct.cpp}
\title{Fast posterior estimation of cell-type fractions and expression with updated reference}
\usage{
fastPost.updated.ct.cpp(bulk_Expr, updated_phi, n.iter = 20, n.core = 1)
}
\arguments{
\item{bulk_Expr}{bulk expression matrix to deconvolute}

\item{updated_phi}{updated reference}

\item{n.iter}{number of iteration}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\description{
Fixed-point implementation of the updated Gibbs sampling step with the updated reference.
     Returns a list of posterior estimation of cell-type fraction.
}
