% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{frac_evalu}
\alias{frac_evalu}
\title{Compare ground truth and estimated cell type fractions}
\usage{
frac_evalu(truth_frac, estimated_frac)
}
\arguments{
\item{truth_frac}{ground truth cell-type fraction with samples in rows and cell-types in columns}

\item{estimated_frac}{cell-type fraction estimation with samples in rows and cell-types in columns}
}
\value{
a list with summary statistics of the correlation analysis
}
\description{
Calculate Pearson correlation between the ground truth and the estimated cell type fractions for each cell type.
   When cell type names in the two fraction matrices do not match exactly, this function automatically identifies the most likely
   correspondences between cell types by finding pairs with the highest pairwise correlation
}
