% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastBP.R
\name{mergeK_adapted}
\alias{mergeK_adapted}
\title{Merge posterior information over cell states}
\usage{
mergeK_adapted(jointPost.obj, map)
}
\arguments{
\item{jointPost.obj}{a jointPost.obj returned by fastPost.ini.cs() function}

\item{map}{a list of the format list(cell.type1=c(cell.stateA, cell.stateB), ...)}
}
\description{
Merge posterior information over cell states
}
\section{Functions}{
\itemize{
\item \code{mergeK_adapted()}: to merge posterior information over cell states within each cell type

}}
