% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{get_Z_array}
\alias{get_Z_array}
\title{Get 3d Z tensor}
\usage{
get_Z_array(InstaPrism_obj, resolution = "ct", n.core = 1)
}
\arguments{
\item{InstaPrism_obj}{An InstaPrism_obj obtained from InstaPrism() function or an InstaPrism_update_obj obtained from InstaPrism_update() function}

\item{resolution}{a character indicating whether to get cell.type specific gene expression or cell.state specific gene expression. Available options include 'ct' or 'cs'. Default = 'ct'}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a gene by cell.type(state) 3d Z array
}
\description{
obtain sampleID by gene by cell.type/cell.state 3d Z array
}
