% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{InstaPrism_reference}
\alias{InstaPrism_reference}
\title{Get InstaPrism built-in reference}
\usage{
InstaPrism_reference(refName, download_locally = F)
}
\arguments{
\item{refName}{Name of the built-in reference. Currently supported refName includes 'BRCA', 'CRC', 'GBM', 'LUAD','OV','RCC','SKCM'}

\item{download_locally}{a logical variable to determine whether to download the reference to the local working directory. Default = F}
}
\value{
an refPhi_cs object that can be pass to the `InstaPrism()` function with `input_type = 'refPhi_cs'`
}
\description{
Load InstaPrism built-in reference and/or download the reference locally
}
\examples{
brca_ref = InstaPrism_reference('BRCA')
}
