% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{merge_Z}
\alias{merge_Z}
\title{Merge posterior information over cell states}
\usage{
merge_Z(Z, map)
}
\arguments{
\item{Z}{a 3d array of cell.state specific expression}

\item{map}{a list of the format list(cell.type1=c(cell.stateA, cell.stateB), ...)}
}
\description{
Function to merge posterior information over cell states within each cell type
}
