% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{reconstruct_Z_ct_initial}
\alias{reconstruct_Z_ct_initial}
\title{Reconstruct cell-type specific expression from an InstaPrism object}
\usage{
reconstruct_Z_ct_initial(
  InstaPrism_obj,
  cell.type.of.interest,
  n.core = 1,
  pseudo.min = 1e-08
)
}
\arguments{
\item{InstaPrism_obj}{an InstaPrism object from InstaPrism() function}

\item{cell.type.of.interest}{cell.type to reconstruct}

\item{n.core}{number of cores to use for parallel programming. Default = 1}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\value{
a gene * sample matrix with cell.type specific expression
}
\description{
Reconstruct cell-type specific expression from an InstaPrism object
}
\examples{
\dontrun{
bulk_expr = read.csv(system.file('extdata','example_bulk.csv',package = 'InstaPrism'))
OV_ref = InstaPrism_reference('OV') # load InstaPrism built-in reference
deconv_res = InstaPrism(bulk_Expr = bulk_expr,refPhi_cs = OV_ref)
# get malignant-specific gene expression
Z_mal = reconstruct_Z_ct_initial(deconv_res,'malignant')
}
}
