% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{reconstruct_Z_ct_updated}
\alias{reconstruct_Z_ct_updated}
\title{Reconstruct cell-type specific expression using the updated reference}
\usage{
reconstruct_Z_ct_updated(
  InstaPrism_updated_obj,
  cell.type.of.interest,
  n.core = 1,
  pseudo.min = 1e-08
)
}
\arguments{
\item{InstaPrism_updated_obj}{an InstaPrism_update object from InstaPrism_update() function}

\item{cell.type.of.interest}{cell.type to reconstruct}

\item{n.core}{number of cores to use for parallel programming. Default = 1}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\value{
a gene * sample matrix with cell.type specific expression
}
\description{
Reconstruct cell-type specific expression using the updated reference
}
\examples{
\dontrun{
bulk_expr = read.csv(system.file('extdata','example_bulk.csv',package = 'InstaPrism'))
OV_ref = InstaPrism_reference('OV') # load InstaPrism built-in reference
deconv_res = InstaPrism(bulk_Expr = bulk_expr,refPhi_cs = OV_ref)
updated_res = InstaPrism_update(deconv_res,bulk_expr,key = 'malignant')
Z_mal_updated = reconstruct_Z_ct_updated(updated_res,'malignant')
}
}
