% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{refPhi-class}
\alias{refPhi-class}
\title{S4 class to store (non-malignant) reference matrix phi or psi (if no malignant cells)}
\description{
S4 class to store (non-malignant) reference matrix phi or psi (if no malignant cells)
}
\section{Slots}{

\describe{
\item{\code{phi}}{a matrix of dimension K*G,
rownames are cell state/type names; colnames are gene IDs/names}

\item{\code{pseudo.min}}{the desired minimum value used to normalize phi.}
}}

