% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{refPrepare}
\alias{refPrepare}
\title{Create reference profile using single cell data}
\usage{
refPrepare(sc_Expr, cell.type.labels, cell.state.labels, pseudo.min = 1e-08)
}
\arguments{
\item{sc_Expr}{single cell expression with genes in rows and cells in columns}

\item{cell.type.labels}{a character vector indicating cell types of each cell}

\item{cell.state.labels}{a character vector indicating cell state of each cell}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\value{
a refPhi_cs object that can be pass to the `InstaPrsim()` function
}
\description{
De novo reference construction using scRNA-seq data
}
\examples{
\dontrun{
library(Biobase)
data("sim.data")
scExpr = exprs(sc.eset)
cell_type_labels = pData(sc.eset)$cellType
refPhi_obj = refPrepare(sc_Expr = scExpr, cell.type.labels = cell_type_labels, cell.state.labels = cell_type_labels)
}
}
