% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{refTumor-class}
\alias{refTumor-class}
\title{S4 class to store reference matrix psi for malignant cells}
\description{
S4 class to store reference matrix psi for malignant cells
}
\section{Slots}{

\describe{
\item{\code{psi_mal}}{a matrix of dimension N*G, to denote updated sample-specific profiles (for malignant cells)
rownames are bulk sample IDs; colnames are gene IDs/names}

\item{\code{psi_env}}{a matrix of dimension (K-1)*G, to denote updated shared profiles (for non-malignant cells)
rownames are non-malignant cell types; colnames are gene IDs/names}

\item{\code{key}}{a character variable to denote the names for malignant cells}

\item{\code{pseudo.min}}{the desired minimum value used to normalize phi.}
}}

