% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{a small simulated dataset to run deconvolution examples}
\format{
An object of class \code{list} of length 4.
}
\usage{
sim.data
}
\description{
a simulated dataset containing  simulate bulk and single-cell expression,
as well as marker genes and true proportions that can be used as an example for deconvolution.
}
\keyword{datasets}
