% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{deconv_performance_plot}
\alias{deconv_performance_plot}
\title{Visualize deconvolution performance}
\usage{
deconv_performance_plot(
  truth_frac,
  estimated_frac,
  title = NULL,
  nrow = 1,
  xlabel = "true fraction",
  ylabel = "estimates"
)
}
\arguments{
\item{truth_frac}{ground truth cell-type fractions with samples in rows and cell-types in columns}

\item{estimated_frac}{cell-type fraction estimates with samples in rows and cell-types in columns}

\item{title}{title for the plot}

\item{nrow}{number of rows in the plot grid. Default = 1}

\item{xlabel}{xlabel of the plot. Default = 'true fraction'}

\item{ylabel}{ylabel of the plot. Default = 'estimates'}
}
\description{
Visualize per-cell type correlation between true fraction and estimates
}
