% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{deconv_performance_plot}
\alias{deconv_performance_plot}
\title{Visualize deconvolution performance}
\usage{
deconv_performance_plot(
  est,
  true,
  title = NULL,
  nrow = 2,
  xlabel = "true fraction",
  ylabel = "estimates"
)
}
\arguments{
\item{est}{cell.type fraction estimation with cell.types in rows and samples in columns}

\item{true}{ground truth cell.type fraction with cell.types in rows and samples in columns}

\item{title}{title for the plot}

\item{nrow}{number of rows in the plot grid}

\item{xlabel}{xlabel of the plot}

\item{ylabel}{ylabel of the plot}
}
\description{
Visualize per-cell type correlation between true fraction and estimates
}
