% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{get_Z_array}
\alias{get_Z_array}
\title{Get 3d Z tensor}
\usage{
get_Z_array(InstaPrism_obj, resolution = "ct", n.core = 1)
}
\arguments{
\item{InstaPrism_obj}{An InstaPrism_obj obtained from InstaPrism() function or an InstaPrism_update_obj obtained from InstaPrism_update() function}

\item{resolution}{a character indicating whether to get cell.type specific gene expression or cell.state specific gene expression. Available options include 'ct' or 'cs'. Default = 'ct'}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a sample by gene by cell.type/cell.state 3d Z array
}
\description{
Obtains a 3-dimensional array of deconvolved gene expression with dimensions sample * gene * cell-type/cell-state.
}
\examples{
\dontrun{
bulk_expr = read.csv(system.file('extdata','example_bulk.csv',package = 'InstaPrism'))
OV_ref = InstaPrism_reference('OV') # load InstaPrism built-in reference
deconv_res = InstaPrism(bulk_Expr = bulk_expr,refPhi_cs = OV_ref)
Z = get_Z_array(deconv_res)
}
}
