% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{InstaPrism}
\alias{InstaPrism}
\title{Run InstaPrism deconvolution}
\usage{
InstaPrism(
  input_type = c("raw", "prism"),
  sc_Expr = NULL,
  bulk_Expr = NULL,
  cell.type.labels = NULL,
  cell.state.labels = NULL,
  outlier.cut = 0.01,
  outlier.fraction = 0.1,
  pseudo.min = 1e-08,
  prismObj = NULL,
  n.iter = 20,
  update = F,
  key = NA,
  optimizer = "MAP",
  opt.control = NULL,
  return.Z = F,
  n.core = 1
)
}
\arguments{
\item{input_type}{either 'raw' or 'prism'. With 'raw', need to specify the following input manually:
sc_Expr, bulk_Expr, cell.type.labels, cell.state.labels.}

\item{sc_Expr}{single cell expression matrix, with genes in rows and cells in columns}

\item{bulk_Expr}{bulk expression to deconvolute, with genes in rows and samples in columns}

\item{cell.type.labels}{a character vector indicating cell types of each cell in sc_Expr}

\item{cell.state.labels}{a character vector indicating cell states of each cell in sc_Expr}

\item{outlier.cut}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.1) of bulk data. Removal of outlier genes will ensure that the inference will not be
dominated by outliers. This parameter denotes the same thing as in new.prism() function from BayesPrism package.}

\item{outlier.fraction}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.1) of bulk data. Removal of outlier genes will ensure that the inference will not be
dominated by outliers. This parameter denotes the same thing as in new.prism() function from BayesPrism package.}

\item{pseudo.min}{the desired min values to replace zero after normalization}

\item{prismObj}{a Prism object, required when input_type='prism'}

\item{n.iter}{number of iterations in the fixed-point implementation of BayesPrism}

\item{update}{a logical variable to denote whether return deconvolution result with the updated theta
This parameter denotes the same thing as in the run.prism() function from BayesPrism. Default=F.}

\item{key}{a charater string to denote the word that corresponds to the malignant cell type, belongs to names(map)
set to NULL if there is no malignant cells in the problem.}

\item{optimizer}{a character string to denote which algorithm to use. Can be either "MAP" or "MLE".
This parameter denotes the same thing as in the updateReference() function from BayesPrism.}

\item{opt.control}{a list containing the parameters to control optimization}

\item{return.Z}{a logical variable determining whether to return cell.state&type specific expression, default=FALSE}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
a list containing cellular fraction estimates, summarized at both cell type level and cell state level
}
\description{
A fast version of BayesPrism deconvolution, takes single cell profiles as prior information
     and returns cellular composition estimation for a bulk expression matrix.
}
\details{
add more detials
}
\references{
ref BayesPrism
}
