% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{InstaPrism}
\alias{InstaPrism}
\title{Run InstaPrism deconvolution}
\usage{
InstaPrism(
  input_type = c("raw", "prism", "refPhi_cs"),
  sc_Expr = NULL,
  bulk_Expr = NULL,
  cell.type.labels = NULL,
  cell.state.labels = NULL,
  filter = TRUE,
  outlier.cut = 0.01,
  outlier.fraction = 0.1,
  pseudo.min = 1e-08,
  prismObj = NULL,
  refPhi_cs = NULL,
  n.iter = NULL,
  verbose = F,
  convergence.plot = F,
  max_n_per_plot = 50,
  n.core = 1
)
}
\arguments{
\item{input_type}{one of the following input_types are allowed: 'raw', 'prism' and 'refPhi_cs'.
With 'raw', need to specify the following input manually: sc_Expr, bulk_Expr, cell.type.labels, cell.state.labels.
With 'prism', need to provide a prism object. With 'refPhi_cs', need to provide refPhi_cs and bulk Expr}

\item{sc_Expr}{single cell expression matrix (without log transformation), with genes in rows and cells in columns, required when input_type = T}

\item{bulk_Expr}{bulk expression to deconvolute (without log transformation), with genes in rows and samples in columns}

\item{cell.type.labels}{a character vector indicating cell types of each cell in sc_Expr}

\item{cell.state.labels}{a character vector indicating cell states of each cell in sc_Expr}

\item{filter}{a logical vatiable to determine whether to filter genes in bulk_Expr or not}

\item{outlier.cut}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.01) of bulk data. Removal of outlier genes will ensure that the inference will not be
dominated by outliers. This parameter denotes the same thing as in new.prism() function from BayesPrism package.}

\item{outlier.fraction}{Filter genes in bulk mixture whose expression fraction is greater than outlier.cut (Default=0.01)
in more than outlier.fraction (Default=0.1) of bulk data. Removal of outlier genes will ensure that the inference will not be
dominated by outliers. This parameter denotes the same thing as in new.prism() function from BayesPrism package.}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}

\item{prismObj}{a Prism object, required when input_type='prism'}

\item{refPhi_cs}{a refPhi_cs object with single cell reference phi information (cell state only), required when input_type='refPhi_cs'}

\item{n.iter}{number of iterations in InstaPrism algorithm. Default = max(100, number of cell.states * 2)}

\item{verbose}{a logical variable determining whether to display convergence status of the model. Default = F}

\item{convergence.plot}{a logical variable determining whether to visualize convergence status for cell types}

\item{max_n_per_plot}{max number of samples to visualize in one convergence plot}

\item{n.core}{number of cores to use for parallel programming. Default = 1}
}
\value{
an InstaPrism object containing posterior information for cell states and cell types
}
\description{
A fast version of BayesPrism deconvolution, takes single cell profiles as prior information
     and returns cellular composition estimation for a bulk expression matrix.
}
