% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{InstaPrism_update}
\alias{InstaPrism_update}
\title{Run InstaPrism update on InstaPrism_elite object}
\usage{
InstaPrism_update(
  InstaPrism_obj,
  bulk_Expr,
  n.iter = 100,
  n.core = 1,
  cell.types.to.update = NULL,
  key = NA,
  keep.phi = "phi.ct",
  pseudo.min = 1e-08
)
}
\arguments{
\item{InstaPrism_obj}{an InstaPrism object from InstaPrism() function}

\item{bulk_Expr}{bulk expression to deconvolute (without log transformation), with genes in rows and samples in columns}

\item{n.iter}{number of iterations in InstaPrism algorithm. Default = 100}

\item{n.core}{number of cores to use for parallel programming. Default = 1}

\item{cell.types.to.update}{non-malignant cell.types to update}

\item{key}{name of the malignant cell type. Upon setting the key parameter, the updated malignant reference will be unique for each individual.
Set to NA if there is no malignant cells in the problem, and the updated reference will be the same for all the individuals}

\item{keep.phi}{either 'phi.ct' or 'phi.cs', denoting whether using phi.ct or phi.cs for cell.types that are not updated}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\value{
an InstaPrism_update object
}
\description{
fast and memory-saving ways to update malignant reference and cell.types of interest
}
