% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{reconstruct_Z_cs_initial}
\alias{reconstruct_Z_cs_initial}
\title{Reconstruct cell.state specific expression from initial.reference}
\usage{
reconstruct_Z_cs_initial(
  InstaPrism_obj,
  cell.state.of.interest,
  pseudo.min = 1e-08
)
}
\arguments{
\item{InstaPrism_obj}{an InstaPrism object from InstaPrism() function}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}

\item{cell.type.of.interest}{cell.type to reconstruct}
}
\value{
a gene * sample matrix with cell.state specific expression
}
\description{
build cell.state specific expression, which is equivalent to Z of Post.ini.ct object from InstaPrism() function
}
