% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{reconstruct_Z_ct_updated}
\alias{reconstruct_Z_ct_updated}
\title{Reconstruct cell.type specific expression using the updated reference}
\usage{
reconstruct_Z_ct_updated(
  InstaPrism_updated_obj,
  cell.type.of.interest,
  key = NA,
  n.core = 1,
  pseudo.min = 1e-08
)
}
\arguments{
\item{InstaPrism_updated_obj}{an InstaPrism_update object from InstaPrism_update() function}

\item{cell.type.of.interest}{cell.type to reconstruct}

\item{key}{name of the malignant cell type. Set to NA if no malignant cells present in the problem}

\item{n.core}{number of cores to use for parallel programming. Default = 1}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\value{
a gene * sample matrix with cell.type specific expression
}
\description{
Reconstruct cell.type specific expression using the updated reference
}
