% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{refPrepare}
\alias{refPrepare}
\title{function to make refPhi from single cell}
\usage{
refPrepare(sc_Expr, cell.type.labels, cell.state.labels, pseudo.min = 1e-08)
}
\arguments{
\item{sc_Expr}{single cell expression with genes in rows and cells in columns}

\item{cell.type.labels}{a character vector indicating cell types of each cell}

\item{cell.state.labels}{a character vector indicating cell state of each cell}

\item{pseudo.min}{pseudo.min value to replace zero for normalization. Default = 1E-8}
}
\description{
function to make refPhi from single cell
}
