% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstaPrism.R
\name{refPrepare}
\alias{refPrepare}
\title{function to make refPhi or refPhi_cs from single cell}
\usage{
refPrepare(
  sc_Expr,
  cell.type.labels,
  cell.state.labels,
  pseudo.min = 1e-08,
  output_class = c("refPhi", "refPhi_cs")
)
}
\arguments{
\item{sc_Expr}{single cell expression with genes in rows and cells in columns}

\item{cell.type.labels}{a character vector indicating cell types of each cell}

\item{cell.state.labels}{a character vector indicating cell state of each cell}

\item{pseudo.min}{the desired min values to replace zero after normalization}
}
\description{
function to make refPhi or refPhi_cs from single cell
}
